requirements.txt
python-dotenv
PyYAML
requests
playwright
anthropic
Pillow
